requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2
django-4-jet==1.0.9
django-appconf==1.0.6
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-compressor==4.5.1
django-js-asset==2.2.0
django-solo==2.3.0
gunicorn==23.0.0
idna==3.10
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
python-dotenv==1.0.1
razorpay==1.4.2
rcssmin==1.1.2
requests==2.32.3
rjsmin==1.2.2
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.3
